+++ lib/td_dd/search/search.ex
@@ -72,10 +72,12 @@ defmodule TdDd.Search do
   end
 
   defp filter_values({name, %{"buckets" => buckets}}) do
-    {name, buckets |> Enum.map(& &1["key"])}
+    {name, Enum.map(buckets, & &1["key"])}
   end
 
   defp filter_values({name, %{"distinct_search" => distinct_search}}) do
     filter_values({name, distinct_search})
   end
+
+  defp filter_values({name, %{"doc_count" => 0}}), do: {name, []}
 end