+++ lib/td_dd_web/views/data_structure_type_view.ex
@@ -1,5 +1,6 @@
 defmodule TdDdWeb.DataStructureTypeView do
   use TdDdWeb, :view
+
   alias TdDdWeb.DataStructureTypeView
 
   def render("index.json", %{data_structure_types: data_structure_types}) do
@@ -10,20 +11,21 @@ defmodule TdDdWeb.DataStructureTypeView do
     %{data: render_one(data_structure_type, DataStructureTypeView, "data_structure_type.json")}
   end
 
-  def render("data_structure_type.json", %{data_structure_type: %{template: template} = data_structure_type}) do
-    %{id: data_structure_type.id,
-      structure_type: data_structure_type.structure_type,
-      translation: data_structure_type.translation,
-      metadata_fields: data_structure_type.metadata_fields,
-      template: %{id: template.id, name: template.name}
-    }
+  def render("data_structure_type.json", %{
+        data_structure_type: %{template: %{} = template} = data_structure_type
+      }) do
+    data_structure_type
+    |> Map.take([:id, :structure_type, :translation, :metadata_fields])
+    |> Map.put(:template, Map.take(template, [:id, :name]))
   end
 
-  def render("data_structure_type.json", %{data_structure_type: data_structure_type}) do
-    %{id: data_structure_type.id,
-      structure_type: data_structure_type.structure_type,
-      translation: data_structure_type.translation,
-      metadata_fields: data_structure_type.metadata_fields,
-      template_id: data_structure_type.template_id}
+  def render("data_structure_type.json", %{data_structure_type: %{} = data_structure_type}) do
+    Map.take(data_structure_type, [
+      :id,
+      :structure_type,
+      :translation,
+      :metadata_fields,
+      :template_id
+    ])
   end
 end