+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+
+- [TD-2486] Template type `domain`
+
 ## [4.9.0] 2020-11-30
 
 ### Changed
@@ -10,7 +14,6 @@
 ### Added
 
 - [TD-3089] Widget and type `copy` on df
-- [TD-2486] Template type `domain`
 
 ### Changed